Refactored challenge and deleted challengeItem class.
